requirements.txt
accelerate==1.10.1
aiodns==3.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==21.3.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.13.5
bitsandbytes==0.47.0
bleach==6.2.0
blinker==1.9.0
Bottleneck==1.4.2
Brotli==1.1.0
brotlicffi==1.0.9.2
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.3.2
click==8.3.0
colorama==0.4.6
comm==0.2.1
contourpy==1.3.1
cycler==0.11.0
datasets==3.3.2
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
et_xmlfile==2.0.0
executing==2.2.1
faiss-cpu==1.12.0
fastapi==0.117.1
fastjsonschema==2.20.0
ffmpy==0.6.1
filelock==3.17.0
fonttools==4.59.2
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
gmpy2==2.2.1
gradio==5.47.2
gradio_client==1.13.3
groovy==0.1.2
groq==0.32.0
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.34.4
idna==3.7
ipykernel==6.30.1
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2023.7.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lark==1.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
narwhals==1.31.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.5
notebook_shim==0.2.4
numexpr==2.11.0
numpy==1.26.4
openpyxl==3.1.5
orjson==3.11.3
overrides==7.4.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.4
pillow==11.1.0
pip==25.2
platformdirs==4.3.7
plotly==6.3.0
prometheus_client==0.21.1
prompt-toolkit==3.0.43
propcache==0.3.1
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.0
pycares==4.10.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.0
PyQt6==6.9.1
PyQt6_sip==13.10.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.30.2
regex==2025.9.1
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.22.3
ruff==0.13.2
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==5.1.1
setuptools==78.1.1
shellingham==1.5.4
sip==6.12.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.0
soupsieve==2.5
stack_data==0.6.3
starlette==0.48.0
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.13.3
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
typer==0.19.2
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
unicodedata2==15.1.0
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.14
win-inet-pton==1.1.0
winloop==0.2.2
xxhash==3.5.0
yarl==1.18.0
